% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_solver.R
\name{adams}
\alias{adams}
\title{Create an Adams solver (adaptive step size)}
\usage{
adams(abs_tol = 1e-10, rel_tol = 1e-10, max_num_steps = 1e+09)
}
\arguments{
\item{abs_tol}{Absolute tolerance.}

\item{rel_tol}{Relative tolerance.}

\item{max_num_steps}{Maximum number of steps.}
}
\description{
Create an Adams solver (adaptive step size)
}
\seealso{
Other ODE solver creation functions: 
\code{\link{bdf}()},
\code{\link{ckrk}()},
\code{\link{rk45}()},
\code{\link{rk4}()}
}
\concept{ODE solver creation functions}
