% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-solver.R
\name{AdaptiveOdeSolver}
\alias{AdaptiveOdeSolver}
\title{An ODE solver with adaptive step size}
\description{
An ODE solver with adaptive step size

An ODE solver with adaptive step size
}
\section{Super class}{
\code{\link[odemodeling:OdeSolver]{odemodeling::OdeSolver}} -> \code{AdaptiveOdeSolver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{abs_tol}}{absolute tolerance}

\item{\code{rel_tol}}{relative tolerance}

\item{\code{max_num_steps}}{maximum number of steps}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdaptiveOdeSolver$new()}}
\item \href{#method-standata}{\code{AdaptiveOdeSolver$standata()}}
\item \href{#method-to_string}{\code{AdaptiveOdeSolver$to_string()}}
\item \href{#method-clone}{\code{AdaptiveOdeSolver$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="odemodeling" data-topic="OdeSolver" data-id="print">}\href{../../odemodeling/html/OdeSolver.html#method-print}{\code{odemodeling::OdeSolver$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveOdeSolver$new(name, abs_tol, rel_tol, max_num_steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{solver name}

\item{\code{abs_tol}}{absolute tolerance}

\item{\code{rel_tol}}{relative tolerance}

\item{\code{max_num_steps}}{maximum number of steps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standata"></a>}}
\if{latex}{\out{\hypertarget{method-standata}{}}}
\subsection{Method \code{standata()}}{
Create Stan data fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveOdeSolver$standata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-to_string}{}}}
\subsection{Method \code{to_string()}}{
String description of the solver.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveOdeSolver$to_string()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveOdeSolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
