% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{compare_odefits}
\alias{compare_odefits}
\alias{max_abs_odesol_diff}
\alias{max_abs_loglik_diff}
\title{Compare two objects that contain ODE solutions and ODE model likelihoods}
\usage{
max_abs_odesol_diff(x, y)

max_abs_loglik_diff(x, y)
}
\arguments{
\item{x}{An object of class \link{OdeModelMCMC} or \link{OdeModelGQ}.}

\item{y}{An object of class \link{OdeModelMCMC} or \link{OdeModelGQ}.}
}
\description{
Compare two objects that contain ODE solutions and ODE model likelihoods
}
\section{Functions}{
\itemize{
\item \code{max_abs_odesol_diff}: Compute maximum absolute difference in
ODE solutions

\item \code{max_abs_loglik_diff}: Compute maximum absolute differences in
log likelihoods
}}

