% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{compute_reliability_metrics}
\alias{compute_reliability_metrics}
\title{Compute reliability metrics}
\usage{
compute_reliability_metrics(mcmc, gq)
}
\arguments{
\item{mcmc}{An object of class \link{OdeModelMCMC}.}

\item{gq}{An object of class \link{OdeModelGQ}.}
}
\value{
A named numeric vector.
}
\description{
Compute reliability metrics
}
