% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_odemodel.R
\name{create_odemodel}
\alias{create_odemodel}
\title{Create an ODE model from parts of 'Stan' code}
\usage{
create_odemodel(
  odefun_data = character(0),
  odefun_pars = character(0),
  odefun_code = "",
  odefun_tdata = character(0),
  tdata_code = "",
  obsmodel_data = character(0),
  obsmodel_pars = character(0),
  priors = character(0),
  loglik_code = "",
  genquant = character(0),
  genquant_code = "",
  verbose = FALSE,
  compile = TRUE,
  ...
)
}
\arguments{
\item{odefun_data}{Additional data needed by the ODE function
(character vector of 'Stan' variable declarations).}

\item{odefun_pars}{Parameters of the ODE function (character vector of
'Stan' parameter declarations).}

\item{odefun_code}{Body of the ODE function (string of 'Stan' code). Must
end in a line that returns the time derivative of the system as a vector.}

\item{odefun_tdata}{Additional transformed data needed by the ODE function
(character vector of 'Stan' variable declarations).}

\item{tdata_code}{String of 'Stan' code that computes the declared
transformed data.}

\item{obsmodel_data}{Additional data needed by the likelihood function
(character vector of 'Stan' variable declarations).}

\item{obsmodel_pars}{Parameters of the likelihood function
(character vector of 'Stan' parameter declarations).}

\item{loglik_code}{Body of the log likelihood function
(string of 'Stan' code). Must end in a line that returns a real number,
that is (proportional to) the likelihood.}

\item{genquant}{Additional generated quantities
(character vector of 'Stan' variable declarations).}

\item{genquant_code}{String of 'Stan' code that computes the declared
generated quantities.}

\item{verbose}{Should this print more information?}

\item{compile}{Should the model be compiled?}

\item{...}{Additional arguments to \code{cmdstanr::write_stan_file()}.}
}
\value{
An object of class \code{OdeModel}.
}
\description{
Create an ODE model from parts of 'Stan' code
}
\seealso{
Other setup functions: 
\code{\link{example_odemodel}()},
\code{\link{generate_stancode}()}
}
\concept{setup functions}
