% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-misc.R
\name{create_solver_conf_list}
\alias{create_solver_conf_list}
\title{Create a list of solver configurations (for adaptive stepsize solvers)}
\usage{
create_solver_conf_list(tols, max_num_steps)
}
\arguments{
\item{tols}{a numeric vector of tolerances (abs tol and rel tol will be same)}

\item{max_num_steps}{maximum number of steps (same in all confs)}
}
\description{
Create a list of solver configurations (for adaptive stepsize solvers)
}
