% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example}
\alias{example}
\title{Run example}
\usage{
example(verbose = FALSE)
}
\arguments{
\item{verbose}{Should this print more information.}
}
\value{
Define return value.
}
\description{
Run example
}
