% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_model}
\alias{example_model}
\title{Create example model}
\usage{
example_model(verbose = FALSE, compile = TRUE)
}
\arguments{
\item{verbose}{Should this print more information?}

\item{compile}{Should the 'Stan' model be compiled?}
}
\value{
A \code{CmdStanModel} object.
}
\description{
Create example model
}
