% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-examples.R
\name{example_ode_model}
\alias{example_ode_model}
\title{Create an example ODE model}
\usage{
example_ode_model(prior_only = FALSE, ...)
}
\arguments{
\item{prior_only}{Create a prior-only version of the model?}

\item{...}{Additional arguments to \code{\link[=ode_model]{ode_model()}}.}
}
\value{
An object of class \code{OdeModel}.
}
\description{
Create an example ODE model
}
\seealso{
Other model constructor functions: 
\code{\link{ode_model}()}
}
\concept{model constructor functions}
