% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stancode.R
\name{generate_stancode_prior}
\alias{generate_stancode_prior}
\title{Generate 'Stan' code for prior model}
\usage{
generate_stancode_prior(
  data = list(),
  tdata = list(),
  params = list(),
  tparams = list()
)
}
\arguments{
\item{data}{A list of \code{StanDeclaration} objects.}

\item{tdata}{A list of \code{StanTransformation} objects.}

\item{params}{A list of \code{StanParameter} objects.}

\item{tparams}{A list of \code{StanTransformation} objects.}
}
\value{
A string defining a complete 'Stan' model.
}
\description{
Generate 'Stan' code for prior model
}
