% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-postproc.R
\name{get_array_draw}
\alias{get_array_draw}
\title{Get an unflattened array draw}
\usage{
get_array_draw(fit, variable, iteration)
}
\arguments{
\item{fit}{A \link[cmdstanr:CmdStanMCMC]{cmdstanr::CmdStanMCMC} object.}

\item{variable}{Name of the array variable}

\item{iteration}{Iteration index of draw to get.}
}
\description{
Get an unflattened array draw
}
