% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odemodeling-package.R
\docType{data}
\name{lynxhare}
\alias{lynxhare}
\title{Hudson Bay Company lynx and hare pelt collection data (1900-1920)}
\format{
A data frame with 21 rows and 3 variables:
\describe{
\item{year}{collection year}
\item{lynx}{number of lynx pelts collected (in 1000's)}
\item{hare}{number of hare pelts collected (in 1000's)}
}
}
\source{
\url{https://www.math.tamu.edu/~phoward/m442/modbasics.pdf},
downloaded 10th Jan, 2021.
}
\usage{
lynxhare
}
\description{
Hudson Bay Company lynx and hare pelt collection data (1900-1920)
}
\keyword{datasets}
