% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-solvers.R
\name{odesolvers}
\alias{odesolvers}
\alias{rk45}
\alias{bdf}
\alias{adams}
\alias{ckrk}
\alias{rk4}
\title{Creating ODE solvers}
\usage{
rk45(abs_tol = 1e-06, rel_tol = 1e-06, max_num_steps = 1e+06)

bdf(abs_tol = 1e-10, rel_tol = 1e-10, max_num_steps = 1e+09)

adams(abs_tol = 1e-10, rel_tol = 1e-10, max_num_steps = 1e+09)

ckrk(abs_tol = 1e-10, rel_tol = 1e-10, max_num_steps = 1e+09)

rk4(num_steps = 1)
}
\arguments{
\item{abs_tol}{Absolute tolerance (only for \link{AdaptiveOdeSolver}s).}

\item{rel_tol}{Relative tolerance (only for \link{AdaptiveOdeSolver}s).}

\item{max_num_steps}{Maximum number of steps between output time
points (only for \link{AdaptiveOdeSolver}s).}

\item{num_steps}{The number of steps between output time points (only for
\link{FixedNumStepsOdeSolver}s).}
}
\description{
These constructors should be used for creating the \code{solver}
argument of \code{\link[=sample_odemodel]{sample_odemodel()}}. Each function here returns
an \link{OdeSolver} which can be either
\itemize{
\item An \link{AdaptiveOdeSolver} which can estimate its own error and
adapts its step size according to given tolerances for the error
\item A \link{FixedNumStepsOdeSolver} which always takes the same number of
steps between consequent output time points
}
}
\section{Functions}{
\itemize{
\item \code{rk45}: Create an RK45 solver (\link{AdaptiveOdeSolver})

\item \code{bdf}: Create a BDF solver (\link{AdaptiveOdeSolver})

\item \code{adams}: Create an Adams solver (\link{AdaptiveOdeSolver})

\item \code{ckrk}: Create a Cash-Karp solver (\link{AdaptiveOdeSolver}).

\item \code{rk4}: Create an RK4 solver (\link{FixedNumStepsOdeSolver})
}}

