% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-plot.R
\name{plot_metric}
\alias{plot_metric}
\alias{plot_pareto_k}
\alias{plot_r_eff}
\alias{plot_mad}
\title{Plotting different metrics}
\usage{
plot_metric(values, name, tols = NULL, num_steps = NULL)

plot_pareto_k(reliability, tols = NULL, num_steps = NULL)

plot_r_eff(reliability, tols = NULL, num_steps = NULL)

plot_mad(reliability, tols = NULL, num_steps = NULL, loglik = FALSE)
}
\arguments{
\item{values}{A numeric vector of metric values.}

\item{name}{Name of metric}

\item{tols}{A numeric vector of tolerances. Must have same length as
\code{values}.}

\item{num_steps}{A numeric vector of step amounts. Must have same length as
\code{values}.}

\item{reliability}{The list returned by the \verb{$reliability()} method of
\link{OdeModelFit} class.}

\item{loglik}{If \code{TRUE}, the maximum absolute difference in log likelihoods
is plotted. Otherwise the maximum absolute difference in ODE solutions
is plotted (default).}
}
\description{
Either \code{tols} or \code{num_steps} has to be \code{NULL}.
}
\section{Functions}{
\itemize{
\item \code{plot_metric}: Plot a generic metric.

\item \code{plot_pareto_k}: Plot pareto-k metric.

\item \code{plot_r_eff}: Plot relative efficiency.

\item \code{plot_mad}: Plot maximum absolute difference.
}}

