% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_solver.R
\name{rk4}
\alias{rk4}
\title{Create an RK4 solver (fixed number of steps)}
\usage{
rk4(num_steps = 1)
}
\arguments{
\item{num_steps}{Number of steps per timepoint interval.}
}
\description{
Create an RK4 solver (fixed number of steps)
}
\seealso{
Other ODE solver creation functions: 
\code{\link{adams}()},
\code{\link{bdf}()},
\code{\link{ckrk}()},
\code{\link{rk45}()}
}
\concept{ODE solver creation functions}
