% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_solver.R
\name{rk45}
\alias{rk45}
\title{Create an RK45 solver (adaptive step size)}
\usage{
rk45(abs_tol = 1e-06, rel_tol = 1e-06, max_num_steps = 1e+06)
}
\arguments{
\item{abs_tol}{Absolute tolerance.}

\item{rel_tol}{Relative tolerance.}

\item{max_num_steps}{Maximum number of steps.}
}
\description{
Create an RK45 solver (adaptive step size)
}
\seealso{
Other ODE solver creation functions: 
\code{\link{adams}()},
\code{\link{bdf}()},
\code{\link{ckrk}()},
\code{\link{rk4}()}
}
\concept{ODE solver creation functions}
