% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-mcmc.R
\name{sample_odemodel_manyconf}
\alias{sample_odemodel_manyconf}
\title{Sample parameters of an ODE model using many different ODE solver
configurations}
\usage{
sample_odemodel_manyconf(
  solvers,
  model,
  t0,
  t,
  data = list(),
  savedir = "results",
  basename = "odemodelfit",
  chains = 4,
  ...
)
}
\arguments{
\item{solvers}{List of ODE solvers (possibly the same solver with
different configurations). See \code{\link[=create_solver_list]{create_solver_list()}} for creating this.}

\item{model}{An object of class \link{OdeModel}.}

\item{t0}{Initial time point.}

\item{t}{Vector of time points.}

\item{data}{Other needed data as a list.}

\item{savedir}{Directory where results are saved.}

\item{basename}{Base name for saved files.}

\item{chains}{Number of MCMC chains.}

\item{...}{Arguments passed to the \verb{$sample()} method of the
underlying \link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel} object.}
}
\value{
A named list.
}
\description{
Sample parameters of an ODE model using many different ODE solver
configurations
}
