% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Create a setup using a 'Stan' file}
\usage{
setup(file)
}
\arguments{
\item{file}{Path to a \code{.stan} file.}
}
\value{
An object of class \code{OdeModelSetup}.
}
\description{
Create a setup using a 'Stan' file
}
