% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-declarations.R
\name{stan_dim}
\alias{stan_dim}
\title{Create a \link{StanDimension} object}
\usage{
stan_dim(name, lower = NULL, upper = NULL)
}
\arguments{
\item{name}{name of the dimension variable}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Create a \link{StanDimension} object
}
\examples{
N <- stan_dim("N")
M <- stan_dim("M", lower = 1, upper = 100)
print(N)
print(M)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_matrix}()},
\code{\link{stan_param}()},
\code{\link{stan_transform}()},
\code{\link{stan_var}()},
\code{\link{stan_vector}()},
\code{\link{stan_vector_array}()}
}
\concept{Stan variable declaration functions}
