% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standecl.R
\name{stan_gq}
\alias{stan_gq}
\title{Create a \code{StanGeneratedQuantity} object}
\usage{
stan_gq(var, code = "")
}
\arguments{
\item{var}{The Stan variable declaration for the quantity.
Must be an object that inherits from \code{StanDeclaration}.}

\item{code}{A string of Stan code that defines how the quantity is
computed. The default is empty string (no definition).}
}
\description{
Create a \code{StanGeneratedQuantity} object
}
\examples{
N <- stan_dim("N")
D <- stan_dim("D")
var <- stan_array("y", dims = list(N, D), type = "int")
code <- "
for(n in 1:N){
  for(d in 1:D) {
    y[n, d] ~ poisson_rng(0.2);
  }
}"
y <- stan_gq(arr, code)
print(y)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_matrix}()},
\code{\link{stan_param}()},
\code{\link{stan_var}()},
\code{\link{stan_vector_array}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
