% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standecl.R
\name{stan_matrix}
\alias{stan_matrix}
\title{Create a \code{StanMatrix} object.}
\usage{
stan_matrix(name, nrow, ncol, lower = NULL, upper = NULL)
}
\arguments{
\item{name}{name of the matrix}

\item{nrow}{number of rows, must be a \code{StanDimension} object}

\item{ncol}{number of columns, must be a \code{StanDimension} object}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Create a \code{StanMatrix} object.
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_var}()},
\code{\link{stan_vector_array}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
