% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standecl.R
\name{stan_param}
\alias{stan_param}
\title{Create a \code{StanParameter} object}
\usage{
stan_param(var, prior_code = "")
}
\arguments{
\item{var}{The Stan variable declaration from which the parameter is
created. Must be an object that inherits from \code{StanDeclaration} and
has a real or vector base type.}

\item{prior_code}{A string of Stan code that defines the prior for the
parameter. The default is empty string (no prior).}
}
\description{
Create a \code{StanParameter} object
}
\examples{
# Scalar parameter
my_par <- stan_param(stan_var("beta"), "beta ~ normal(0, 1)")
print(my_par)

# Vector parameter
my_vec <- stan_vector("alpha", stan_dim("D"), lower = 0)
my_par <- stan_param(my_vec)
print(my_par)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_matrix}()},
\code{\link{stan_var}()},
\code{\link{stan_vector_array}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
