% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standecl.R
\name{stan_var}
\alias{stan_var}
\title{Create a \code{StanVariable} object.}
\usage{
stan_var(name, type = "real", lower = NULL, upper = NULL)
}
\arguments{
\item{name}{name of the dimension variable}

\item{type}{variable type (real or int)}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Create a \code{StanVariable} object.
}
\examples{
x <- stan_var("x", type = "real", lower = 0)
y <- stan_var("y", type = "int")
print(x)
print(y)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_matrix}()},
\code{\link{stan_vector_array}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
