% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_standecl.R
\name{stan_vector}
\alias{stan_vector}
\title{Create a \link{StanVector} object.}
\usage{
stan_vector(name, length, lower = NULL, upper = NULL)
}
\arguments{
\item{name}{name of the vector}

\item{length}{length of the vector, must be a \link{StanDimension} object}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Create a \link{StanVector} object.
}
\examples{
x <- stan_vector("x", length = stan_dim("N", lower = 1), lower = 0)
print(x)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_matrix}()},
\code{\link{stan_param}()},
\code{\link{stan_transform}()},
\code{\link{stan_var}()},
\code{\link{stan_vector_array}()}
}
\concept{Stan variable declaration functions}
