% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standecl.R
\name{stan_vector_array}
\alias{stan_vector_array}
\title{Create a \code{StanVectorArray} object.}
\usage{
stan_vector_array(name, dims, length, lower = NULL, upper = NULL)
}
\arguments{
\item{name}{name of the vector array}

\item{dims}{list of array dimensions, must be a list of
\code{StanDimension} objects}

\item{length}{length of the vector, must be a \code{StanDimension} object}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Create a \code{StanVectorArray} object.
}
\examples{
N <- stan_dim("N")
D <- stan_dim("D")
vec_arr <- stan_vector_array("y", dims = list(N), length = D)
print(vec_arr)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_matrix}()},
\code{\link{stan_param}()},
\code{\link{stan_transform}()},
\code{\link{stan_var}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
