% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanArray}
\alias{StanArray}
\title{A Stan array}
\description{
A Stan array

A Stan array
}
\section{Super class}{
\code{\link[odemodeling:StanDeclaration]{odemodeling::StanDeclaration}} -> \code{StanArray}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dims}}{list of array dimensions}

\item{\code{type}}{base type of the array}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanArray-new}{\code{StanArray$new()}}
\item \href{#method-StanArray-declaration}{\code{StanArray$declaration()}}
\item \href{#method-StanArray-signature}{\code{StanArray$signature()}}
\item \href{#method-StanArray-can_be_made_parameter}{\code{StanArray$can_be_made_parameter()}}
\item \href{#method-StanArray-get_dims}{\code{StanArray$get_dims()}}
\item \href{#method-StanArray-clone}{\code{StanArray$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="print"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-print'><code>odemodeling::StanDeclaration$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanArray-new}{}}}
\subsection{Method \code{new()}}{
Create a \link{StanArray} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$new(name, dims, type = "real", lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the array}

\item{\code{dims}}{list of array dimensions, must be a list of
\link{StanDimension} objects}

\item{\code{type}}{base type of the array}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanArray-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-StanArray-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a declaration for the array
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$declaration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanArray-signature"></a>}}
\if{latex}{\out{\hypertarget{method-StanArray-signature}{}}}
\subsection{Method \code{signature()}}{
The variable when used in function signature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$signature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanArray-can_be_made_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-StanArray-can_be_made_parameter}{}}}
\subsection{Method \code{can_be_made_parameter()}}{
Can the object be made into a parameter?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$can_be_made_parameter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanArray-get_dims"></a>}}
\if{latex}{\out{\hypertarget{method-StanArray-get_dims}{}}}
\subsection{Method \code{get_dims()}}{
Get all declared dimensions related to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$get_dims()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanArray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanArray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
