% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanMatrix}
\alias{StanMatrix}
\title{A Stan matrix}
\description{
A Stan matrix

A Stan matrix
}
\section{Super class}{
\code{\link[odemodeling:StanDeclaration]{odemodeling::StanDeclaration}} -> \code{StanMatrix}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nrow}}{number of rows}

\item{\code{ncol}}{number of columns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanMatrix-new}{\code{StanMatrix$new()}}
\item \href{#method-StanMatrix-declaration}{\code{StanMatrix$declaration()}}
\item \href{#method-StanMatrix-signature}{\code{StanMatrix$signature()}}
\item \href{#method-StanMatrix-can_be_made_parameter}{\code{StanMatrix$can_be_made_parameter()}}
\item \href{#method-StanMatrix-get_dims}{\code{StanMatrix$get_dims()}}
\item \href{#method-StanMatrix-clone}{\code{StanMatrix$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="print"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-print'><code>odemodeling::StanDeclaration$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanMatrix-new}{}}}
\subsection{Method \code{new()}}{
Create a \link{StanMatrix} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMatrix$new(name, nrow, ncol, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the matrix}

\item{\code{nrow}}{number of rows, must be a \link{StanDimension} object}

\item{\code{ncol}}{number of columns, must be a \link{StanDimension} object}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMatrix-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-StanMatrix-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a declaration for the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMatrix$declaration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMatrix-signature"></a>}}
\if{latex}{\out{\hypertarget{method-StanMatrix-signature}{}}}
\subsection{Method \code{signature()}}{
The variable when used in function signature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMatrix$signature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMatrix-can_be_made_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-StanMatrix-can_be_made_parameter}{}}}
\subsection{Method \code{can_be_made_parameter()}}{
Can the object be made into a parameter?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMatrix$can_be_made_parameter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMatrix-get_dims"></a>}}
\if{latex}{\out{\hypertarget{method-StanMatrix-get_dims}{}}}
\subsection{Method \code{get_dims()}}{
Get all declared dimensions related to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMatrix$get_dims()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
