% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanVector}
\alias{StanVector}
\title{A Stan vector}
\description{
A Stan vector

A Stan vector
}
\section{Super class}{
\code{\link[odemodeling:StanDeclaration]{odemodeling::StanDeclaration}} -> \code{StanVector}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{length}}{length of the vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanVector-new}{\code{StanVector$new()}}
\item \href{#method-StanVector-declaration}{\code{StanVector$declaration()}}
\item \href{#method-StanVector-signature}{\code{StanVector$signature()}}
\item \href{#method-StanVector-can_be_made_parameter}{\code{StanVector$can_be_made_parameter()}}
\item \href{#method-StanVector-get_dims}{\code{StanVector$get_dims()}}
\item \href{#method-StanVector-clone}{\code{StanVector$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="print"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-print'><code>odemodeling::StanDeclaration$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVector-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanVector-new}{}}}
\subsection{Method \code{new()}}{
Create a \link{StanVector} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVector$new(name, length, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the vector}

\item{\code{length}}{length of the vector, must be a \link{StanDimension} object}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVector-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-StanVector-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a declaration for the vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVector$declaration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVector-signature"></a>}}
\if{latex}{\out{\hypertarget{method-StanVector-signature}{}}}
\subsection{Method \code{signature()}}{
The variable when used in function signature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVector$signature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVector-can_be_made_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-StanVector-can_be_made_parameter}{}}}
\subsection{Method \code{can_be_made_parameter()}}{
Can the object be made into a parameter?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVector$can_be_made_parameter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVector-get_dims"></a>}}
\if{latex}{\out{\hypertarget{method-StanVector-get_dims}{}}}
\subsection{Method \code{get_dims()}}{
Get all declared dimensions related to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVector$get_dims()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanVector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
