% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_odemodel.R
\name{create_odemodel}
\alias{create_odemodel}
\title{Create an ODE model from parts of 'Stan' code}
\usage{
create_odemodel(
  odefun_add_args,
  odefun_body,
  loglik_add_args,
  loglik_body,
  data = "",
  middle_blocks = "",
  prior = "",
  genquant_decl = "",
  genquant = "",
  verbose = FALSE,
  compile = TRUE,
  ...
)
}
\arguments{
\item{odefun_add_args}{ODE function additional arguments (vector of
Stan code strings).}

\item{odefun_body}{ODE function body (Stan code string).}

\item{loglik_add_args}{Log-likelihood function additional arguments
(vector of Stan code strings).}

\item{loglik_body}{Log-likelihood function bod (Stan code string).}

\item{data}{Code to declare data (Stan code string).}

\item{middle_blocks}{The \verb{transformed data}, \code{parameters}, and
\verb{transformed parameters} blocks (Stan code string).}

\item{prior}{Code that defines the prior for parameters (Stan code string).}

\item{genquant_decl}{Declarations of generated quantities (Stan code string).}

\item{genquant}{Code that computes generated quantities (Stan code string).}

\item{verbose}{Should this print more information?}

\item{compile}{Should the model be compiled?}

\item{...}{Additional arguments to \code{cmdstanr::write_stan_file()}.}
}
\value{
An object of class \code{OdeModel}.
}
\description{
Create an ODE model from parts of 'Stan' code
}
\seealso{
Other setup functions: 
\code{\link{example_odemodel}()},
\code{\link{generate_stancode}()}
}
\concept{setup functions}
