% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_odemodel.R
\name{create_odemodel}
\alias{create_odemodel}
\title{Create an ODE model}
\usage{
create_odemodel(
  N,
  odefun_vars = list(),
  odefun_body = "",
  odefun_init = NULL,
  loglik_vars = list(),
  loglik_body = "",
  other_vars = list(),
  verbose = FALSE,
  compile = TRUE,
  sig_figs = 12
)
}
\arguments{
\item{N}{A \link{StanDimension} variable describing the number of time points.}

\item{odefun_vars}{Data and parameters needed by the ODE function. Must be a
list of \link{StanDeclaration}, \link{StanParameter}, or \link{StanTransformation} objects.
These will be defined in Stan model code blocks}

\item{odefun_body}{ODE function body (Stan code string).}

\item{odefun_init}{Initial value for ODE system at t0.
Has to be a \link{StanVector}, or alternatively a \link{StanParameter} or a
\link{StanTransformation} with \link{StanVector} base declaration.}

\item{loglik_vars}{Data and parameters needed by the log likelihood
function.}

\item{loglik_body}{Log likelihood function body (Stan code string).}

\item{other_vars}{Other variables.}

\item{verbose}{Should this print more information?}

\item{compile}{Should the model be compiled?}

\item{sig_figs}{Number of significant figures to use in all Stan i/o.}
}
\value{
An object of class \link{OdeModel}.
}
\description{
Generate model given the declarations of variables,
parameters, functions etc. The arguments \code{odefun_vars},
\code{loglik_vars} and \code{other_vars} must be lists whose elements must have one
of the following three types:
\itemize{
\item \link{StanDeclaration} - can be created using \code{\link[=stan_var]{stan_var()}},
\code{\link[=stan_vector]{stan_vector()}}, \code{\link[=stan_array]{stan_array()}} etc.
\item \link{StanParameter} - can be created using \code{\link[=stan_param]{stan_param()}}
\item \link{StanTransformation} - can be created using \code{\link[=stan_transform]{stan_transform()}}
}

These will go to different blocks of the 'Stan' model code so that
\itemize{
\item \link{StanDeclaration}s go to \code{data}
\item \link{StanParameter}s go to \code{parameters}
\item \link{StanTransformation}s with origin \code{"data"} go to
\verb{transformed data}
\item \link{StanTransformation}s with origin \code{"param"} go to
\verb{transformed parameters}
\item \link{StanTransformation}s with origin \code{"model"} go to
\verb{generated quantities}
}
}
\seealso{
Other setup functions: 
\code{\link{example_odemodel}()}
}
\concept{setup functions}
