% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_odemodel}
\alias{example_odemodel}
\title{Create an example ODE model}
\usage{
example_odemodel(...)
}
\arguments{
\item{...}{Additional arguments to \code{\link{create_odemodel}}.}
}
\value{
An object of class \code{OdeModel}.
}
\description{
Create an example ODE model
}
\seealso{
Other setup functions: 
\code{\link{create_odemodel}()}
}
\concept{setup functions}
