% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-solver.R
\name{FixedNumStepsOdeSolver}
\alias{FixedNumStepsOdeSolver}
\title{An ODE solver with fixed number of steps}
\description{
An ODE solver with fixed number of steps

An ODE solver with fixed number of steps
}
\section{Super class}{
\code{\link[odemodeling:OdeSolver]{odemodeling::OdeSolver}} -> \code{FixedNumStepsOdeSolver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{num_steps}}{maximum number of steps per time point interval}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedNumStepsOdeSolver-new}{\code{FixedNumStepsOdeSolver$new()}}
\item \href{#method-FixedNumStepsOdeSolver-standata}{\code{FixedNumStepsOdeSolver$standata()}}
\item \href{#method-FixedNumStepsOdeSolver-to_string}{\code{FixedNumStepsOdeSolver$to_string()}}
\item \href{#method-FixedNumStepsOdeSolver-clone}{\code{FixedNumStepsOdeSolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeSolver" data-id="print"><a href='../../odemodeling/html/OdeSolver.html#method-OdeSolver-print'><code>odemodeling::OdeSolver$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedNumStepsOdeSolver-new"></a>}}
\if{latex}{\out{\hypertarget{method-FixedNumStepsOdeSolver-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedNumStepsOdeSolver$new(name, num_steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{solver name}

\item{\code{num_steps}}{number of steps per time point interval}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedNumStepsOdeSolver-standata"></a>}}
\if{latex}{\out{\hypertarget{method-FixedNumStepsOdeSolver-standata}{}}}
\subsection{Method \code{standata()}}{
Create Stan data fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedNumStepsOdeSolver$standata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedNumStepsOdeSolver-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-FixedNumStepsOdeSolver-to_string}{}}}
\subsection{Method \code{to_string()}}{
String description of the solver.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedNumStepsOdeSolver$to_string()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedNumStepsOdeSolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FixedNumStepsOdeSolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedNumStepsOdeSolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
