% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-solver.R
\name{FixedNumStepsOdeSolver}
\alias{FixedNumStepsOdeSolver}
\title{An ODE solver with fixed number of steps}
\description{
An ODE solver with fixed number of steps

An ODE solver with fixed number of steps
}
\section{Super class}{
\code{\link[odemodeling:OdeSolver]{odemodeling::OdeSolver}} -> \code{FixedNumStepsOdeSolver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{num_steps}}{maximum number of steps per time point interval}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FixedNumStepsOdeSolver$new()}}
\item \href{#method-clone}{\code{FixedNumStepsOdeSolver$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedNumStepsOdeSolver$new(name, num_steps)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{solver name}

\item{\code{num_steps}}{number of steps per time point interval}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedNumStepsOdeSolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
