% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stancode.R
\name{generate_stancode}
\alias{generate_stancode}
\title{Generate 'Stan' model code}
\usage{
generate_stancode(
  N,
  odefun_data = list(),
  odefun_tdata = list(),
  odefun_params = list(),
  odefun_tparams = list(),
  odefun_body = "",
  loglik_data = list(),
  loglik_tdata = list(),
  loglik_params = list(),
  loglik_tparams = list(),
  loglik_body = "",
  gqs = list()
)
}
\arguments{
\item{N}{A dimension.}

\item{odefun_data}{Data needed by the ODE function. List of
\code{StanDeclaration} objects.}

\item{odefun_tdata}{Transformed data needed by the ODE function.
A list of \code{StanTransformation} objects.}

\item{odefun_params}{Parameters of the ODE function. A list of
\code{StanParameter} objects.}

\item{odefun_tparams}{Transformations for parameters of the ODE
function. A list of \code{StanTransformation} objects.}

\item{odefun_body}{ODE function body (Stan code string).}

\item{loglik_data}{Data needed by the log likelihood function. List of
\code{StanDeclaration} objects.}

\item{loglik_tdata}{Transformed data needed by the log likelihood function.
A list of \code{StanTransformation} objects.}

\item{loglik_params}{Parameters of the log likelihood function. A list of
\code{StanParameter} objects.}

\item{loglik_tparams}{Transformations for parameters of the log likelihood
function. A list of \code{StanTransformation} objects.}

\item{loglik_body}{Log likelihood function body (Stan code string).}

\item{gqs}{Generated quantities. A list of \code{StanTransformation} objects.}
}
\value{
Model code as a string.
}
\description{
Generate 'Stan' model code given the declarations of variables,
parameters, functions etc.
}
\seealso{
Other setup functions: 
\code{\link{create_odemodel}()},
\code{\link{example_odemodel}()}
}
\concept{setup functions}
