% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stancode.R
\name{generate_stancode}
\alias{generate_stancode}
\title{Generate full 'Stan' code given missing parts}
\usage{
generate_stancode(
  odefun_add_args,
  odefun_body,
  loglik_add_args,
  loglik_body,
  data = "",
  middle_blocks = "",
  prior = "",
  genquant_decl = "",
  genquant = "",
  ...
)
}
\arguments{
\item{odefun_add_args}{ODE function additional arguments (vector of
Stan code strings).}

\item{odefun_body}{ODE function body (Stan code string).}

\item{loglik_add_args}{Log-likelihood function additional arguments
(vector of Stan code strings).}

\item{loglik_body}{Log-likelihood function bod (Stan code string).}

\item{data}{Code to declare data (Stan code string).}

\item{middle_blocks}{The \verb{transformed data}, \code{parameters}, and
\verb{transformed parameters} blocks (Stan code string).}

\item{prior}{Code that defines the prior for parameters (Stan code string).}

\item{genquant_decl}{Declarations of generated quantities (Stan code string).}

\item{genquant}{Code that computes generated quantities (Stan code string).}

\item{...}{Additional arguments to \code{cmdstanr::write_stan_file()}.}
}
\value{
Model code as a string and path to file where the model
is saved.
}
\description{
Generate full 'Stan' code given missing parts
}
