% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-model.R
\name{OdeModel}
\alias{OdeModel}
\title{An ODE model (R6 class)}
\description{
An ODE model (R6 class)

An ODE model (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prior}}{An object of class \code{StanModelWithCode}.}

\item{\code{posterior}}{An object of class \code{StanModelWithCode}.}

\item{\code{odetuner_version}}{of the package used to create the model}

\item{\code{sig_figs}}{Number of significant figures to use everywhere.}

\item{\code{t_dim}}{A \code{StanDimension} of the time points array.}

\item{\code{ode_dim}}{A \code{StanDimension} of the ODE system.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OdeModel$new()}}
\item \href{#method-assert_files_exist}{\code{OdeModel$assert_files_exist()}}
\item \href{#method-reinit}{\code{OdeModel$reinit()}}
\item \href{#method-print}{\code{OdeModel$print()}}
\item \href{#method-clone}{\code{OdeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{OdeModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$new(prior, posterior, sig_figs, t_dim, ode_dim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{An object of class \code{StanModelWithCode}.}

\item{\code{posterior}}{An object of class \code{StanModelWithCode}.}

\item{\code{sig_figs}}{Number of significant figures to use in all Stan i/o.}

\item{\code{t_dim}}{Time points vector dimension variable.}

\item{\code{ode_dim}}{ODE system dimension variable.}

\item{\code{compile}}{Should the models be compiled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_files_exist"></a>}}
\if{latex}{\out{\hypertarget{method-assert_files_exist}{}}}
\subsection{Method \code{assert_files_exist()}}{
Check that the Stan models have been initialized correctly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$assert_files_exist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reinit"></a>}}
\if{latex}{\out{\hypertarget{method-reinit}{}}}
\subsection{Method \code{reinit()}}{
(Re)initialize the Stan models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$reinit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print information about the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
