% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{OdeModel}
\alias{OdeModel}
\title{An ODE model.}
\description{
An ODE model.

An ODE model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of model.}

\item{\code{stanmodel}}{An object of class \code{CmdStanModel}.}

\item{\code{datasim}}{Can the model simulate data?}

\item{\code{stancode}}{Full 'Stan' code as a string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OdeModel$new()}}
\item \href{#method-print}{\code{OdeModel$print()}}
\item \href{#method-sample_prior}{\code{OdeModel$sample_prior()}}
\item \href{#method-sample_posterior}{\code{OdeModel$sample_posterior()}}
\item \href{#method-sample_posterior_many}{\code{OdeModel$sample_posterior_many()}}
\item \href{#method-plot}{\code{OdeModel$plot()}}
\item \href{#method-set_data}{\code{OdeModel$set_data()}}
\item \href{#method-set_init}{\code{OdeModel$set_init()}}
\item \href{#method-set_hmc_initial_step_size}{\code{OdeModel$set_hmc_initial_step_size()}}
\item \href{#method-clone}{\code{OdeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{OdeModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$new(stanmodel, datasim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stanmodel}}{An object of class \code{CmdStanModel}.}

\item{\code{...}}{Arguments passed to \verb{$sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print information about the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Sample from parameter prior (no ODE solving)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_prior(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \verb{$sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-sample_posterior}{}}}
\subsection{Method \code{sample_posterior()}}{
Sample from parameter posterior
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_posterior(solver_args, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \verb{$sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_posterior_many"></a>}}
\if{latex}{\out{\hypertarget{method-sample_posterior_many}{}}}
\subsection{Method \code{sample_posterior_many()}}{
Sample from parameter posterior using several different ODE tolereances
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_posterior_many(
  res_dir,
  idx,
  tols,
  max_num_steps,
  chains = 4,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \verb{$sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$plot(fit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$set_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_init"></a>}}
\if{latex}{\out{\hypertarget{method-set_init}{}}}
\subsection{Method \code{set_init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$set_init(init)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hmc_initial_step_size"></a>}}
\if{latex}{\out{\hypertarget{method-set_hmc_initial_step_size}{}}}
\subsection{Method \code{set_hmc_initial_step_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$set_hmc_initial_step_size(step_size)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
