% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-odemodel.R
\name{OdeModel}
\alias{OdeModel}
\title{An ODE model (R6 class)}
\description{
An ODE model (R6 class)

An ODE model (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of model.}

\item{\code{stanmodel}}{An object of class \code{CmdStanModel}.}

\item{\code{datasim}}{Can the model simulate data?}

\item{\code{stancode}}{Full 'Stan' code as a string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OdeModel$new()}}
\item \href{#method-print}{\code{OdeModel$print()}}
\item \href{#method-sample_prior}{\code{OdeModel$sample_prior()}}
\item \href{#method-sample_posterior}{\code{OdeModel$sample_posterior()}}
\item \href{#method-sample_posterior_many}{\code{OdeModel$sample_posterior_many()}}
\item \href{#method-clone}{\code{OdeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{OdeModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$new(stancode, datasim, compile, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stancode}}{Full 'Stan' code as a string.}

\item{\code{datasim}}{Can the model simulate data?}

\item{\code{compile}}{Should the model be compiled?}

\item{\code{...}}{Arguments passed to \code{cmdstanr::write_stan_file()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print information about the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Sample from parameter prior (no ODE solving)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_prior(t0 = 0, t = c(1, 2, 3), D = 1, add_data = list(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t0}}{initial time point}

\item{\code{t}}{data time points}

\item{\code{D}}{ODE system dimension}

\item{\code{...}}{Arguments passed to \verb{$sample()}.}

\item{\code{add}}{data list of additional data needed to run the model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-sample_posterior}{}}}
\subsection{Method \code{sample_posterior()}}{
Sample from parameter posterior
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_posterior(solver_args, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \verb{$sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_posterior_many"></a>}}
\if{latex}{\out{\hypertarget{method-sample_posterior_many}{}}}
\subsection{Method \code{sample_posterior_many()}}{
Sample from parameter posterior using several different ODE tolerances
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_posterior_many(
  res_dir,
  idx,
  tols,
  max_num_steps,
  chains = 4,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \verb{$sample()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
