% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-model.R
\name{OdeModel}
\alias{OdeModel}
\title{An ODE model (R6 class)}
\description{
An ODE model (R6 class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{has_likelihood}}{Is there a likelihood function?}

\item{\code{stanmodel}}{An object of class \code{StanModelWithCode}.}

\item{\code{odetuner_version}}{of the package used to create the model}

\item{\code{sig_figs}}{Number of significant figures to use everywhere.}

\item{\code{t_dim}}{A \code{StanDimension} of the time points array.}

\item{\code{ode_dim}}{A \code{StanDimension} of the ODE system.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OdeModel$new()}}
\item \href{#method-assert_stanfile_exists}{\code{OdeModel$assert_stanfile_exists()}}
\item \href{#method-reinit}{\code{OdeModel$reinit()}}
\item \href{#method-print}{\code{OdeModel$print()}}
\item \href{#method-code}{\code{OdeModel$code()}}
\item \href{#method-sample}{\code{OdeModel$sample()}}
\item \href{#method-sample_manyconf}{\code{OdeModel$sample_manyconf()}}
\item \href{#method-clone}{\code{OdeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{OdeModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$new(has_likelihood, stanmodel, sig_figs, t_dim, ode_dim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{has_likelihood}}{Is there a likelihood function?}

\item{\code{stanmodel}}{An object of class \code{StanModelWithCode}
(will be deepcopied)..}

\item{\code{sig_figs}}{Number of significant figures to use in all Stan i/o.}

\item{\code{t_dim}}{Time points vector dimension variable
(will be deepcopied).}

\item{\code{ode_dim}}{ODE system dimension variable (will be deepcopied).}

\item{\code{compile}}{Should the models be compiled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assert_stanfile_exists"></a>}}
\if{latex}{\out{\hypertarget{method-assert_stanfile_exists}{}}}
\subsection{Method \code{assert_stanfile_exists()}}{
Check that the Stan model has been initialized correctly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$assert_stanfile_exists()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reinit"></a>}}
\if{latex}{\out{\hypertarget{method-reinit}{}}}
\subsection{Method \code{reinit()}}{
(Re)initialize the Stan model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$reinit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print information about the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-code"></a>}}
\if{latex}{\out{\hypertarget{method-code}{}}}
\subsection{Method \code{code()}}{
Get the Stan code of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Sample parameters of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample(t0, t, data = list(), solver = rk45(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t0}}{Initial time point.}

\item{\code{t}}{Vector of time points.}

\item{\code{data}}{Other needed data as a list.}

\item{\code{solver}}{An object of class \link{OdeSolver}.}

\item{\code{...}}{Arguments passed to the \verb{$sample()} method of the
underlying \link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \link{OdeModelMCMC}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_manyconf"></a>}}
\if{latex}{\out{\hypertarget{method-sample_manyconf}{}}}
\subsection{Method \code{sample_manyconf()}}{
Sample parameters of the ODE model using many different ODE solver
configurations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$sample_manyconf(
  solvers,
  t0,
  t,
  data = list(),
  savedir = "results",
  basename = "odemodelfit",
  chains = 4,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solvers}}{List of ODE solvers (possibly the same solver with
different configurations). See \code{\link[=rk45_list]{rk45_list()}} and \code{\link[=bdf_list]{bdf_list()}}
for creating this.}

\item{\code{t0}}{Initial time point.}

\item{\code{t}}{Vector of time points.}

\item{\code{data}}{Other needed data as a list.}

\item{\code{savedir}}{Directory where results are saved.}

\item{\code{basename}}{Base name for saved files.}

\item{\code{chains}}{Number of MCMC chains.}

\item{\code{...}}{Additional arguments passed to the \verb{$sample()} method of the
underlying \link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
