% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-fit.R
\name{OdeModelGQ}
\alias{OdeModelGQ}
\title{An ODE model GQ fit (R6 class)}
\description{
Used for holding the output of the \verb{$gqs()}
method of the \link{OdeModel} and \link{OdeModelMCMC} class. Users are not meant to
instantiate objects of this class directly.
}
\seealso{
For more useful methods, see the methods inherited from
\link{OdeModelFit}.

Other model fit classes: 
\code{\link{OdeModelFit}},
\code{\link{OdeModelMCMC}}
}
\concept{model fit classes}
\section{Super class}{
\code{\link[odemodeling:OdeModelFit]{odemodeling::OdeModelFit}} -> \code{OdeModelGQ}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OdeModelGQ-print}{\code{OdeModelGQ$print()}}
\item \href{#method-OdeModelGQ-clone}{\code{OdeModelGQ$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="cmdstan_seed"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-cmdstan_seed'><code>odemodeling::OdeModelFit$cmdstan_seed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="cmdstan_version"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-cmdstan_version'><code>odemodeling::OdeModelFit$cmdstan_version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="dim"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-dim'><code>odemodeling::OdeModelFit$dim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="dim_odesol"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-dim_odesol'><code>odemodeling::OdeModelFit$dim_odesol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="draws"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-draws'><code>odemodeling::OdeModelFit$draws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="extract_odesol"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-extract_odesol'><code>odemodeling::OdeModelFit$extract_odesol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="extract_odesol_df"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-extract_odesol_df'><code>odemodeling::OdeModelFit$extract_odesol_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="extract_odesol_df_dist"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-extract_odesol_df_dist'><code>odemodeling::OdeModelFit$extract_odesol_df_dist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="extract_odesol_quantile"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-extract_odesol_quantile'><code>odemodeling::OdeModelFit$extract_odesol_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="extract_unflattened"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-extract_unflattened'><code>odemodeling::OdeModelFit$extract_unflattened()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="extract_y0"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-extract_y0'><code>odemodeling::OdeModelFit$extract_y0()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="get_t"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-get_t'><code>odemodeling::OdeModelFit$get_t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="get_t0"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-get_t0'><code>odemodeling::OdeModelFit$get_t0()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="info"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-info'><code>odemodeling::OdeModelFit$info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="initialize"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-initialize'><code>odemodeling::OdeModelFit$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="loglik"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-loglik'><code>odemodeling::OdeModelFit$loglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="nchains"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-nchains'><code>odemodeling::OdeModelFit$nchains()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="ndraws"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-ndraws'><code>odemodeling::OdeModelFit$ndraws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="niterations"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-niterations'><code>odemodeling::OdeModelFit$niterations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="plot_odesol"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-plot_odesol'><code>odemodeling::OdeModelFit$plot_odesol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="plot_odesol_dist"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-plot_odesol_dist'><code>odemodeling::OdeModelFit$plot_odesol_dist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="summary"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-summary'><code>odemodeling::OdeModelFit$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="OdeModelFit" data-id="time"><a href='../../odemodeling/html/OdeModelFit.html#method-OdeModelFit-time'><code>odemodeling::OdeModelFit$time()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeModelGQ-print"></a>}}
\if{latex}{\out{\hypertarget{method-OdeModelGQ-print}{}}}
\subsection{Method \code{print()}}{
Print information about the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModelGQ$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OdeModelGQ-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OdeModelGQ-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeModelGQ$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
