% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odemodeling-package.R
\docType{package}
\name{odemodeling-package}
\alias{odemodeling-package}
\alias{odemodeling}
\title{The 'odemodeling' package}
\description{
Building and fitting models with different numerical
solvers using Stan. Includes efficient tuning and validation of
numerical solvers using Pareto-smoothed importance sampling (PSIS) and
its diagnostics.
}
\section{Creating a model}{

\itemize{
\item Declare model data, parameters, and other variables using
\code{\link[=stan_array]{stan_array()}}, \code{\link[=stan_dim]{stan_dim()}}, \code{\link[=stan_matrix]{stan_matrix()}}, \code{\link[=stan_param]{stan_param()}},
\code{\link[=stan_transform]{stan_transform()}}, \code{\link[=stan_vector_array]{stan_vector_array()}}, and \code{\link[=stan_vector]{stan_vector()}}.
\item Create an \link{OdeModel} model using \code{\link[=ode_model]{ode_model()}}.
}
}

\section{Using different ODE solvers}{

\itemize{
\item See \code{\link{odesolvers}} for constructors that create objects
of class \link{OdeSolver}.
}
}

\section{Fitting a model}{

\itemize{
\item Sample the posterior or prior distribution of the model parameters,
and generate corresponding ODE solutions using the \verb{$sample()} method
of the \link{OdeModel} class.
\item See methods of the \link{OdeModelMCMC} class for studying the returned
object.
}
}

\section{Additional simulation of ODE systems}{

\itemize{
\item See the \verb{$simulate()} method of the \link{OdeModelMCMC} class.
\item See methods of the \link{OdeModelGQ} class for studying the returned
object.
}
}

