% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-solver.R
\name{OdeSolver}
\alias{OdeSolver}
\title{An ODE solver (an abstract base class)}
\description{
An ODE solver (an abstract base class)

An ODE solver (an abstract base class)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name of the solver}

\item{\code{stan_number}}{integer that indexes the solver in the Stan code}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OdeSolver$new()}}
\item \href{#method-standata}{\code{OdeSolver$standata()}}
\item \href{#method-to_string}{\code{OdeSolver$to_string()}}
\item \href{#method-print}{\code{OdeSolver$print()}}
\item \href{#method-clone}{\code{OdeSolver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\code{OdeSolver} is an abstract class that can't be initialized.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeSolver$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standata"></a>}}
\if{latex}{\out{\hypertarget{method-standata}{}}}
\subsection{Method \code{standata()}}{
Create Stan data fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeSolver$standata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-to_string}{}}}
\subsection{Method \code{to_string()}}{
String description of the solver.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeSolver$to_string()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print info about the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeSolver$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OdeSolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
