% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-solvers.R
\name{odesolvers_lists}
\alias{odesolvers_lists}
\alias{rk45_list}
\alias{bdf_list}
\title{Creating lists of ODE solvers}
\usage{
rk45_list(tols, max_num_steps = 1e+06)

bdf_list(tols, max_num_steps = 1e+06)
}
\arguments{
\item{tols}{A vector of length \code{L} of tolerance values.}

\item{max_num_steps}{Maximum number of steps between output time
points (one number, same for all solvers in the output list).}
}
\description{
These constructors can be used for creating the \code{solver}
argument of \code{\link[=sample_odemodel_manyconf]{sample_odemodel_manyconf()}}. Each function here returns
a list of \link{OdeSolver}s.
}
\section{Functions}{
\itemize{
\item \code{rk45_list}: Create a list of RK45 solvers with different
tolerances.

\item \code{bdf_list}: Create a list of BDF solvers with
different tolerances.
}}

