% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odetuner-package.R
\docType{package}
\name{odetuner-package}
\alias{odetuner-package}
\alias{odetuner}
\title{The 'odetuner' package}
\description{
Efficient tuning and validation of solver tolerances in Bayesian
ODE models using Pareto-smoothed importance sampling (PSIS) and
its diagnostics.
}
