% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odetuner-package.R
\docType{package}
\name{odetuner-package}
\alias{odetuner-package}
\alias{odetuner}
\title{The 'odetuner' package}
\description{
Efficient tuning and validation of solver tolerances in Bayesian
ODE models using Pareto-smoothed importance sampling (PSIS) and
its diagnostics.
}
\section{Main functions}{

\itemize{
\item Create a model using \link{create_odemodel}.
\item Create the input needed for \link{create_odemodel} using
\code{\link[=stan_array]{stan_array()}}, \code{\link[=stan_dim]{stan_dim()}}, \code{\link[=stan_matrix]{stan_matrix()}}, \code{\link[=stan_param]{stan_param()}},
\code{\link[=stan_transform]{stan_transform()}}, \code{\link[=stan_vector_array]{stan_vector_array()}}, and \code{\link[=stan_vector]{stan_vector()}}.
\item Sample the posterior or prior distribution of the model parameters,
and generate corresponding ODE solutions using \link{sample_odemodel}.
}
}

