% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_odemodel.R
\name{sample_odemodel}
\alias{sample_odemodel}
\title{Sample parameters  of an ODE model}
\usage{
sample_odemodel(
  model,
  t0,
  t,
  data = list(),
  solver = "rk45",
  solver_conf = NULL,
  prior_only = FALSE,
  ...
)
}
\arguments{
\item{model}{An object of class \code{OdeModel}.}

\item{t0}{Initial time point.}

\item{t}{Vector of time points.}

\item{data}{Other needed data as a list.}

\item{solver}{ODE solver name.}

\item{solver_conf}{List of ODE solver configuration arguments.}

\item{prior_only}{Sample only from the prior? If this is true, ODE
solves are done only in generated quantities, and there is no need to
compute gradients for the solutions.}

\item{...}{Arguments passed to the \verb{$sample()} method of the
underlying \code{CmdStanModel} objects.}
}
\description{
Sample parameters  of an ODE model
}
