% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-sample_odemodel.R
\name{sample_odemodel}
\alias{sample_odemodel}
\title{Sample parameters  of an ODE model}
\usage{
sample_odemodel(
  model,
  t0,
  t,
  data = list(),
  solver = "rk45",
  solver_conf = NULL,
  ...
)
}
\arguments{
\item{model}{An object of class \link{OdeModel}.}

\item{t0}{Initial time point.}

\item{t}{Vector of time points.}

\item{data}{Other needed data as a list.}

\item{solver}{ODE solver name.}

\item{solver_conf}{List of ODE solver configuration arguments.}

\item{...}{Arguments passed to the \verb{$sample()} method of the
underlying \link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel} objects.}
}
\description{
Sample parameters  of an ODE model
}
