% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-sample_odemodel.R
\name{sample_odemodel_manyconf}
\alias{sample_odemodel_manyconf}
\title{Sample parameters  of an ODE model using many different ODE solver
configurations}
\usage{
sample_odemodel_manyconf(
  model,
  t0,
  t,
  data = list(),
  solver = "rk45",
  solver_confs = list(),
  savedir = "results",
  basename = "odemodelfit",
  chains = 4,
  ...
)
}
\arguments{
\item{model}{An object of class \link{OdeModel}.}

\item{t0}{Initial time point.}

\item{t}{Vector of time points.}

\item{data}{Other needed data as a list.}

\item{solver}{ODE solver name.}

\item{solver_confs}{List of ODE solver configurations. See
\code{\link[=create_solver_conf_list]{create_solver_conf_list()}} for creating this.}

\item{savedir}{Directory where results are saved.}

\item{basename}{Base name for saved files.}

\item{chains}{Number of MCMC chains.}

\item{...}{Arguments passed to the \verb{$sample()} method of the
underlying \link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel} objects.}
}
\description{
Sample parameters  of an ODE model using many different ODE solver
configurations
}
