% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_odemodel.R
\name{sample_odemodel_manyconf}
\alias{sample_odemodel_manyconf}
\title{Sample parameters  of an ODE model using many different ODE solver
configurations}
\usage{
sample_odemodel_manyconf(
  model,
  t0,
  t,
  data = list(),
  solver = "rk45",
  solver_confs = list(),
  prior_only = FALSE,
  savedir = "results",
  basename = "out",
  chains = 4,
  ...
)
}
\arguments{
\item{model}{An object of class \code{OdeModel}.}

\item{t0}{Initial time point.}

\item{t}{Vector of time points.}

\item{data}{Other needed data as a list.}

\item{solver}{ODE solver name.}

\item{solver_confs}{List of ODE solver configurations.}

\item{prior_only}{Sample only from the prior? If this is true, ODE
solves are done only in generated quantities, and there is no need to
compute gradients for the solutions.}

\item{savedir}{Directory where results are saved.}

\item{basename}{Base name for saved files.}

\item{chains}{Number of MCMC chains.}

\item{...}{Arguments passed to the \verb{$sample()} method of the
underlying \code{CmdStanModel} objects.}
}
\description{
Sample parameters  of an ODE model using many different ODE solver
configurations
}
