% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_standecl.R
\name{stan_matrix}
\alias{stan_matrix}
\title{Create a \link{StanMatrix} object.}
\usage{
stan_matrix(name, nrow, ncol, lower = NULL, upper = NULL)
}
\arguments{
\item{name}{name of the matrix}

\item{nrow}{number of rows, must be a \link{StanDimension} object}

\item{ncol}{number of columns, must be a \link{StanDimension} object}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Create a \link{StanMatrix} object.
}
\examples{
N <- stan_dim("N")
M <- stan_dim("M", lower = 1, upper = 100)
my_mat <- stan_matrix("A", nrow = N, ncol = M, lower = 0, upper = 3.2)
print(my_mat)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_param}()},
\code{\link{stan_transform}()},
\code{\link{stan_var}()},
\code{\link{stan_vector_array}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
