% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-declarations.R
\name{stan_transform}
\alias{stan_transform}
\title{Create a \link{StanTransformation} object}
\usage{
stan_transform(decl, origin = "model", code = "")
}
\arguments{
\item{decl}{The Stan variable declaration for the quantity.
Must be an object that inherits from \link{StanDeclaration}.}

\item{origin}{Must be either \code{"data"}, \code{"parameters"}, or \code{"model"}.
These correspond to the \verb{transformed data}, \verb{transformed parameters}, and
\verb{generated quantities} blocks, respectively.}

\item{code}{A string of Stan code that defines how the quantity is
computed from other variables/parameters.
The default is empty string (no definition). If \code{code}
doesn't include any assignments (with \code{=}), it is prepended
with \code{paste0(decl$name, " = ")}. In this case also a semicolon is
added to the end if it is missing.}
}
\description{
Create a \link{StanTransformation} object
}
\examples{
N <- stan_dim("N")
D <- stan_dim("D")
decl <- stan_array("y", dims = list(N, D), type = "int")
code <- "
for(n in 1:N) {
  for(d in 1:D) {
    y[n, d] ~ poisson_rng(0.2);
  }
}"
y <- stan_transform(decl, code = code)
print(y)
}
\seealso{
Other Stan variable declaration functions: 
\code{\link{stan_array}()},
\code{\link{stan_dim}()},
\code{\link{stan_matrix}()},
\code{\link{stan_param}()},
\code{\link{stan_var}()},
\code{\link{stan_vector_array}()},
\code{\link{stan_vector}()}
}
\concept{Stan variable declaration functions}
