% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanArray}
\alias{StanArray}
\title{A Stan array}
\description{
A Stan array

A Stan array
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name of the array}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}

\item{\code{dims}}{list of array dimensions}

\item{\code{type}}{base type of the array}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StanArray$new()}}
\item \href{#method-declaration}{\code{StanArray$declaration()}}
\item \href{#method-print}{\code{StanArray$print()}}
\item \href{#method-clone}{\code{StanArray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{StanArray} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$new(name, dims, type = "real", lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the array}

\item{\code{dims}}{list of array dimensions, must be a list of
\code{StanDimension} objects}

\item{\code{type}}{base type of the array}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a declaration for the array
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$declaration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
