% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanDimension}
\alias{StanDimension}
\title{A dimension for a Stan vector or array}
\description{
A dimension for a Stan vector or array

A dimension for a Stan vector or array
}
\section{Super class}{
\code{\link[odemodeling:StanDeclaration]{odemodeling::StanDeclaration}} -> \code{StanDimension}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanDimension-new}{\code{StanDimension$new()}}
\item \href{#method-StanDimension-declaration}{\code{StanDimension$declaration()}}
\item \href{#method-StanDimension-signature}{\code{StanDimension$signature()}}
\item \href{#method-StanDimension-get_dims}{\code{StanDimension$get_dims()}}
\item \href{#method-StanDimension-clone}{\code{StanDimension$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="can_be_made_parameter"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-can_be_made_parameter'><code>odemodeling::StanDeclaration$can_be_made_parameter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="print"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-print'><code>odemodeling::StanDeclaration$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanDimension-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanDimension-new}{}}}
\subsection{Method \code{new()}}{
Create a \link{StanDimension} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanDimension$new(name, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the dimension variable}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanDimension-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-StanDimension-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a data block declaration for the dimension.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanDimension$declaration()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanDimension-signature"></a>}}
\if{latex}{\out{\hypertarget{method-StanDimension-signature}{}}}
\subsection{Method \code{signature()}}{
The variable when used in function signature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanDimension$signature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanDimension-get_dims"></a>}}
\if{latex}{\out{\hypertarget{method-StanDimension-get_dims}{}}}
\subsection{Method \code{get_dims()}}{
Get all declared dimensions related to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanDimension$get_dims()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanDimension-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanDimension-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanDimension$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
