% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanVariable}
\alias{StanVariable}
\title{A Stan variable}
\description{
A Stan variable

A Stan variable
}
\section{Super class}{
\code{\link[odemodeling:StanDeclaration]{odemodeling::StanDeclaration}} -> \code{StanVariable}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{type of the variable}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanVariable-new}{\code{StanVariable$new()}}
\item \href{#method-StanVariable-declaration}{\code{StanVariable$declaration()}}
\item \href{#method-StanVariable-signature}{\code{StanVariable$signature()}}
\item \href{#method-StanVariable-can_be_made_parameter}{\code{StanVariable$can_be_made_parameter()}}
\item \href{#method-StanVariable-clone}{\code{StanVariable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="get_dims"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-get_dims'><code>odemodeling::StanDeclaration$get_dims()</code></a></span></li>
<li><span class="pkg-link" data-pkg="odemodeling" data-topic="StanDeclaration" data-id="print"><a href='../../odemodeling/html/StanDeclaration.html#method-StanDeclaration-print'><code>odemodeling::StanDeclaration$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVariable-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanVariable-new}{}}}
\subsection{Method \code{new()}}{
Create a \link{StanVariable} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVariable$new(name, type = "real", lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the dimension variable}

\item{\code{type}}{variable type (real or int)}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVariable-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-StanVariable-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a declaration for the variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVariable$declaration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVariable-signature"></a>}}
\if{latex}{\out{\hypertarget{method-StanVariable-signature}{}}}
\subsection{Method \code{signature()}}{
The variable when used in function signature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVariable$signature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVariable-can_be_made_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-StanVariable-can_be_made_parameter}{}}}
\subsection{Method \code{can_be_made_parameter()}}{
Can the object be made into a parameter?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVariable$can_be_made_parameter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
