% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-standecl.R
\name{StanVectorArray}
\alias{StanVectorArray}
\title{A Stan array of vectors}
\description{
A Stan array of vectors

A Stan array of vectors
}
\section{Super class}{
\code{\link[odetuner:StanDeclaration]{odetuner::StanDeclaration}} -> \code{StanVectorArray}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dims}}{list of array dimensions}

\item{\code{length}}{length of the vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StanVectorArray$new()}}
\item \href{#method-declaration}{\code{StanVectorArray$declaration()}}
\item \href{#method-signature}{\code{StanVectorArray$signature()}}
\item \href{#method-can_be_made_parameter}{\code{StanVectorArray$can_be_made_parameter()}}
\item \href{#method-get_dims}{\code{StanVectorArray$get_dims()}}
\item \href{#method-clone}{\code{StanVectorArray$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{StanVectorArray} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVectorArray$new(name, dims, length, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the vector array}

\item{\code{dims}}{list of array dimensions, must be a list of
\code{StanDimension} objects}

\item{\code{length}}{length of the vector, must be a \code{StanDimension} object}

\item{\code{lower}}{lower bound}

\item{\code{upper}}{upper bound}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-declaration"></a>}}
\if{latex}{\out{\hypertarget{method-declaration}{}}}
\subsection{Method \code{declaration()}}{
Create a declaration for the array of vectors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVectorArray$declaration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-signature"></a>}}
\if{latex}{\out{\hypertarget{method-signature}{}}}
\subsection{Method \code{signature()}}{
The variable when used in function signature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVectorArray$signature()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-can_be_made_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-can_be_made_parameter}{}}}
\subsection{Method \code{can_be_made_parameter()}}{
Can the object be made into a parameter?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVectorArray$can_be_made_parameter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dims"></a>}}
\if{latex}{\out{\hypertarget{method-get_dims}{}}}
\subsection{Method \code{get_dims()}}{
Get all declared dimensions related to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVectorArray$get_dims()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanVectorArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
